+++ jwst/photom/photom.py
@@ -10,6 +10,8 @@ from stdatamodels.jwst.datamodels import dqflags
 
 from .. lib.wcs_utils import get_wavelengths
 from . import miri_mrs
+from . import miri_imager
+from jwst.datamodels import MirImgPhotomModel
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
@@ -468,7 +470,32 @@ class DataSet():
                 row = find_row(ftab.phot_table, fields_to_match)
                 if row is None:
                     return
-            self.photom_io(ftab.phot_table[row])
+                self.photom_io(ftab.phot_table[row])
+                
+            # Check if reference file contains the coefficients for the time-dependent correction of the PHOTOM value
+            try:
+                ftab.getarray_noinit("timecoeff")
+                log.info("Applying the time-dependent correction to the PHOTOM value.")
+
+                mid_time = self.input.meta.exposure.mid_time
+                photom_corr = miri_imager.time_corr_photom(ftab.timecoeff[row], mid_time)
+
+                data = np.array(
+                    [(self.filter, self.subarray, ftab.phot_table[row]['photmjsr']+photom_corr, ftab.phot_table[row]['uncertainty'])],
+                    dtype=[
+                        ("filter", "O"),
+                        ("subarray", "O"),
+                        ("photmjsr", "<f4"),
+                        ("uncertainty", "<f4")
+                        ],
+                )
+                fftab = MirImgPhotomModel(phot_table=data)
+                self.photom_io(fftab.phot_table[0])
+            except:
+                # No time-dependent correction is applied
+                log.info(" Skipping MIRI imager time correction. Extension not found in the reference file.")
+                self.photom_io(ftab.phot_table[row])
+
         # MRS detectors
         elif self.detector == 'MIRIFUSHORT' or self.detector == 'MIRIFULONG':