+++ CHANGES.rst
@@ -57,6 +57,11 @@ imprint
 - Updated the logging to report which imprint image is being subtracted from the
   science image. [#8041]
 
+photom
+--------
+
+- Added time-dependent correction for MIRI Imager data [#, spacetelescope/stdatamodels#235]
+
 ramp_fitting
 ------------